fix(ui/worker): rename fg worker file name
